HIP19 initial commit Radon Tech LLC
Novel, but heavier than helium gas, radon project falls on the ground and hits the network with vertical solutions adding value to the ecosystem.

Radon Tech LLC https://radon.tech/ is an IoT company with scope in LATAM that was born from a network of experienced Argentine professionals led by Jose Vigil and encouraged by the novel cryptocurrency proposition applied to wireless. Proposes breaking the mold with an out-of-the-box hotspot (aka HS) through design, modularity, custom hardware, and a multi-interfacing system for improving the user experience.