Radon Tech LLC HIP19 Hotspot Manufacturers Submit.
A novel, but heavier than helium gas, the radon project falls on the ground and hits the network with vertical solutions adding value to the ecosystem.

Radon Tech LLC is an IoT company with scope in LATAM that was born from a network of experienced Argentine professionals led by Jose Vigil and encouraged by the novel cryptocurrency proposition applied to wireless. Proposes breaking the mold with an out-of-the-box hotspot (aka HS) through design, modularity, custom hardware, and a multi-interfacing system for improving the user experience.